Add baud rate support patch for telnet per RFC2217
Idea is based on an implementation obtained from EMC Isilon, but is a
bit more complete, in particular this version checks baud rates for
sanity, sets the command handlers in a compatible manner, and attempts
to get rid of duplicated code in telnet/telnetd with baud rates (the
whole thing should be cleaned up probably because there's little value
in lugging around BSD 4.2/4.3 compat in this day in age).

This is only a minimal implementation of the items in RFC2217; items
like SET-DATASIZE, SET-PARITY and SET-STOPSIZE need to be handled as
well in order to be fully RFC compliant.

Obtained-from: EMC Isilon
Signed-off-by: Garrett Cooper <yanegomi@gmail.com>