Remove dependency on datetime module
After doing some fidgeting around trying to figure out why snapshots
weren't being deleted as expected, it turns out that the
`datetime.relativedelta` module is buggy (somehow) and is replacing the
fields in the `datetime` objects instead of incrementing/decrementing
them.

I haven't tracked down the bug yet, but obviously this is a no-go for
me.

In the meantime, approximate the number of weeks in a month and year in
order to workaround the fact that `datetime.timedelta` lacks these
fields (lame).