Provide an example for using `[unittest.]mock`
The point of this example is to provide a self-documented test suite
for implementing a fizzbuzz API which reports errors to an external
endpoint via the `FizzBuzzer.report_failure(..)` public API. Given
that the calling API (in this example) isn't guaranteed to be
implemented, simulate that case. While here, add an additional hook
for making `FizzBuzzer.report_failure(..)` flaky, to illustrate the
power of not relying on internal APIs in unit tests.

The tests can be triggered using tox.