Clean up merge-git-branches.sh
A few things I noted about recent changes where I'm basically trying
to merge against multiple remotes with a single clone is the need to
map many things to one remote but not others, and the need to not
merge everything to a single supported remote.

Example:

Say branch A was on remote X and branch B was on remote Y. Previously
it would mirror all checked out branches on remote X to remote X and
vice versa, which is far from desirable--hence the need to run
`git push -f X Y`, in a 1:1 manner, for each appropriate remote.

I [still] think this approach is clumsy and while it does better,
this script is approaching the point of needing to be integrated
into better tools/frameworks that integrate with git to get a greater
degree of introspection, instead of having to define a funky naming
scheme where topics could be mapped back to a single master branch.