Export unsummited list data to a gpx file
This file uses BeautifulSoup(4) to take a list with ascent data on it,
dumps all of the unsummited entries, grabs their GPS coordinates, then
outputs the aggregate of the coordinates to a single .GPX file.

This helped me save copious amounts of time which I previously spent
making maps of unsummited peaks on CalTopo using tedious manual data
entry.