Add helper script for flattening out output from clang-cpp
This change introduces a python script whose sole purpose is to parse
output from clang-cpp and provide meaningful information about what is
contained within the file itself. It is particularly helpful with header
files.

The impetus for this work was to avoid the tedious task of having to
manually parse OpenSSL headers so it would be easier to determine what
APIs/constants may have changed from an interface perspective between
versions and what things have remained the same.